✨ 테스트 결과 그룹화를 Judge가 담당한다
