+++ core/test-registry/types.ts
@@ -4,3 +4,6 @@ export type TResultStatus =
   | "flaky"
   | "not_executed"
   | "manual_only";
+
+export type TTestSuiteId = string; // TC-x
+export type TTestCaseId = string; // TC-x.x.x
